+++ contracts/contracts/protocol/v1/ZAccountsRegistry.sol
@@ -229,7 +229,7 @@ contract ZAccountsRegistry is
     function activateZAccount(
         uint256[] calldata inputs,
         SnarkProof calldata proof,
-        uint256 cachedForestRootIndexAndTaxiEnabler,
+        uint32 cachedForestRootIndexAndTaxiEnabler,
         uint96 paymasterCompensation,
         bytes memory privateMessages
     ) external returns (uint256 utxoBusQueuePos) {
@@ -452,7 +452,7 @@ contract ZAccountsRegistry is
     function _createZAccountUTXO(
         uint256[] calldata inputs,
         SnarkProof calldata proof,
-        uint256 cachedForestRootIndexAndUtxosCarrier,
+        uint32 cachedForestRootIndexAndUtxosCarrier,
         uint96 paymasterCompensation,
         bytes memory privateMessages
     ) private returns (uint256 utxoBusQueuePos) {
